Transform git-depends into a generic dispatcher.
Dispatches git commands onto dependency repository. Eliminates the need
for git-depends-diff. Echos the name of the repository before each
command. The first argument is a regex filter on the name(s). Be careful
executing destructive commands on multiple dependencies. In the future,
perhaps option parsing could be added to require a --force option when
multiple dependencies match or something.